Moving AcquireLockResponse to transpor package
