+++ src/main/java/org/opensearch/jobscheduler/transport/AcquireLockResponse.java
@@ -6,7 +6,7 @@
  * this file be licensed under the Apache-2.0 license or a
  * compatible open source license.
  */
-package org.opensearch.jobscheduler.model;
+package org.opensearch.jobscheduler.transport;
 
 import java.io.IOException;
